requirements.txt
Pillow>=9.0.0
PyQt5>=5.15.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
rembg>=2.0.50
onnxruntime-gpu>=1.15.0
